requirements.txt
streamlit
python-dotenv
redis
pinecone
pinecone-plugin-assistant